test(fleet-comms): track migration v3 rollout
X-Agent: codex/6078-acp-conversation
