+++ tests/test_runtime_api.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 import json
 import sqlite3
 from datetime import UTC, datetime, timedelta
+from pathlib import Path
 
 from fastapi.testclient import TestClient
 
@@ -12,6 +13,7 @@ import scripts.api.runtime_router as runtime_router
 from scripts.api.main import app
 
 client = TestClient(app, raise_server_exceptions=False)
+DASHBOARDS = Path(__file__).resolve().parents[1] / "dashboards"
 
 
 def _iso(dt: datetime) -> str:
@@ -438,6 +440,46 @@ def test_acp_summary_prefers_terminal_wall_duration_and_token_total(tmp_path, mo
 
     assert summary["total_duration_ms"] == 250
     assert summary["total_tokens"] == 5
+
+
+def test_runtime_page_has_a_separate_read_only_active_acp_timeline():
+    html = (DASHBOARDS / "runtime.html").read_text(encoding="utf-8")
+    acp_panel = html[html.index('id="acp-heading"') : html.index('id="acpx-heading"')]
+
+    assert "Active ACP Conversations" in acp_panel
+    assert "/api/runtime/acp/conversations?limit=12" in html
+    assert "/api/runtime/acp/conversations/${encodeURIComponent(summary.conversation_id)}" in html
+    assert "Loading active conversations..." in acp_panel
+    assert "No active ACP conversations." in html
+    assert "Conversation storage is unavailable." in html
+    assert "Malformed conversation data" in html
+    assert "Root</div><div class=\"acp-lane\">Codex</div><div class=\"acp-lane\">Grok" in html
+    assert "round ${event.round}" in html
+    assert "seq ${event.sequence}" in html
+    assert "Duplicate suppressed" in html
+    assert "Ended: ${displayLabel(conversation.termination_reason)}" in html
+    assert "acpStateClass" in html
+    assert "acpStateLabel" in html
+    for label in ["Queued", "Running", "Succeeded", "Partial", "Cancelled", "Failed"]:
+        assert label in html
+    for state in ["COMPLETE", "PARTIAL_COMPLETE", "CANCELLED", "FAILED", "CREATED"]:
+        assert state in html
+
+    for prohibited in [
+        "<form",
+        "acp-send",
+        "acp-post",
+        "acp-retry",
+        "acp-cancel",
+        "acp-route",
+        "acp-review",
+        "acp-config",
+    ]:
+        assert prohibited not in acp_panel
+
+    assert html.index('id="acp-heading"') < html.index('id="acpx-heading"')
+    assert "ACPX Shadow Transport" in html
+    assert "ACPX evidence is observational only." in html
 
 
 def test_acp_termination_reason_allowlists_budget_and_deadline_events():
